chore(server): rename req and res values
