+++ packages/aspida-server/apis/users/_userId@number/@controller.ts
@@ -3,5 +3,5 @@ import { Methods } from '.'
 import { Values } from './$values'
 
 export default createController<Methods, Values>({
-  get: async ({ params }) => ({ status: 200, resBody: { id: params.userId, name: 'bbb' } })
+  get: async ({ params }) => ({ status: 200, body: { id: params.userId, name: 'bbb' } })
 })
